+++ controller/rq_hover.mli
@@ -6,3 +6,20 @@
 (************************************************************************)
 
 val hover : Request.position
+
+open Fleche
+
+module Handler : sig
+  (** Returns [Some markdown] if there is some hover to match *)
+  type 'node h =
+    contents:Contents.t -> point:int * int -> node:'node -> string option
+
+  type t =
+    | MaybeNode : Doc.Node.t option h -> t
+    | WithNode : Doc.Node.t h -> t
+end
+
+(** Register a hover plugin *)
+module Register : sig
+  val add : Handler.t -> unit
+end