+++ CHANGES.md
@@ -24,7 +24,8 @@
    (@ejgallego, #561, reported by @Zimmi48, fixes #548)
  - Provide basic notation information on hover. This is intended for
    people to build their own more refined notation feedback systems
-   (@ejgallego, #)
+   (@ejgallego, #562)
+ - Hover request can now be extended by plugins (@ejgallego, #562)
 
 # coq-lsp 0.1.7: Just-in-time
 -----------------------------